chore: reorganize project file structure with package
