+++ src/main/kotlin/ecommerce/dao/ProductRepository.kt
@@ -1,5 +1,6 @@
-package ecommerce.api
+package ecommerce.dao
 
+import ecommerce.model.Product
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.RowMapper
 import org.springframework.stereotype.Repository
