+++ src/main/kotlin/ecommerce/view/ProductViewController.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.view
 
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.GetMapping
