+++ src/test/kotlin/ecommerce/api/ProductControllerTest.kt
@@ -1,5 +1,7 @@
 package ecommerce.api
 
+import ecommerce.dao.ProductRepository
+import ecommerce.model.Product
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
